Rename dns submodule to protocol module
